+++ app/src/main/java/woowacourse/movie/result/ReservationResultContract.kt
@@ -1,0 +1,15 @@
+package woowacourse.movie.result
+
+import woowacourse.movie.domain.Reservation
+
+interface ReservationResultContract {
+    interface View {
+        fun bindReservation(reservation: Reservation)
+    }
+
+    interface Presenter {
+        fun setUpReservation(reservation: Reservation)
+
+        fun showReservation()
+    }
+}
