+++ app/src/main/java/woowacourse/movie/result/ReservationResultPresenter.kt
@@ -1,0 +1,17 @@
+package woowacourse.movie.result
+
+import woowacourse.movie.domain.Reservation
+
+class ReservationResultPresenter(
+    private val view: ReservationResultContract.View,
+) : ReservationResultContract.Presenter {
+    private lateinit var reservation: Reservation
+
+    override fun setUpReservation(reservation: Reservation) {
+        this.reservation = reservation
+    }
+
+    override fun showReservation() {
+        view.bindReservation(reservation)
+    }
+}
