+++ app/src/main/res/layout/activity_reservation_result.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ReservationResultActivity">
+    tools:context=".result.ReservationResultActivity">
 
     <TextView
         android:id="@+id/tv_cancel_guide"
